requirements_for_test.txt
torch_ema
tqdm
huggingface_hub
transformers
tensorboardX
accelerate
typed-argument-parser
autocorrect
nltk
rembg
openai
trimesh
torchvision
imageio
sentencepiece
protobuf
rich
matplotlib
scipy
numpy 
pandas